Able to Navigate to CurrencyNavigationFragment
