+++ app/build.gradle
@@ -3,6 +3,7 @@ apply plugin: 'com.google.firebase.firebase-perf'
 apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: 'androidx.navigation.safeargs.kotlin'
 
 android {
     compileSdkVersion 28
@@ -62,10 +63,6 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
 
     annotationProcessor "org.parceler:parceler:1.1.12"
-}
-
-repositories {
-    mavenCentral()
 }
 
 apply plugin: 'com.google.gms.google-services'
